Develop

+++ functions.sh
@@ -4,13 +4,14 @@ function parse_arg()
     b=${a//"-"/}
     c=${b//"+"/}
     d=${c//"."/}
-    if grep "Created package for $1$" $LOG
+    echo $d
+    if grep "Created package for $1$" >/dev/null 2>&1 $LOG
     then
-	echo "$1 alredy builded"
+	echo "$1 alredy builded ($LOG)"
     else
         export $d=1
     fi
-    echo $d
+
 }
 
 function clear_flags()
@@ -87,13 +88,19 @@ function pkgpack()
     tar -zcvf ../${PKGNAME}_${PKGVERSION}_${PKGBUILD}_${ABI}.tar.gz .
 }
 
+function clean_pkgdir()
+{
+    [ -z "$KEEP_PKGDIR"  ] && rm -r $PKGDIR || true
+    unset $KEEP_PKGDIR
+}
+
 function package_make()
 {
     export PKGNAME=$1
     export PKGVERSION=$2
     export PKGBUILD=$3
     PKGDIR=$CWD/_packages/$ABI/$PKGNAME
-    rm -r $PKGDIR || true
+    clean_pkgdir
     mkdir -p ${PKGDIR}${SHORTPREFIX}/var/lib/packages
     export DESTDIR=$PKGDIR
     make install || true
@@ -113,7 +120,7 @@ function package_waf()
     export PKGVERSION=$2
     export PKGBUILD=$3
     PKGDIR=$CWD/_packages/$ABI/$PKGNAME
-    rm -r $PKGDIR || true
+    clean_pkgdir
     mkdir -p ${PKGDIR}${SHORTPREFIX}/var/lib/packages
     ./waf -v install --destdir=$PKGDIR || true
     pkgstrip
@@ -131,7 +138,7 @@ function package_b2()
     export PKGVERSION=$2
     export PKGBUILD=$3
     PKGDIR=$CWD/_packages/$ABI/$PKGNAME
-    rm -r $PKGDIR || true
+    clean_pkgdir
     mkdir -p ${PKGDIR}${SHORTPREFIX}/var/lib/packages
     OLDPATH=$PATH
     export PATH="$TOOLCHAIN/bin:$PATH"
@@ -155,7 +162,7 @@ function package_meson()
     export PKGVERSION=$2
     export PKGBUILD=$3
     PKGDIR=$CWD/_packages/$ABI/$PKGNAME
-    rm -r $PKGDIR || true
+    clean_pkgdir
     mkdir -p ${PKGDIR}${SHORTPREFIX}/var/lib/packages
     export DESTDIR=$PKGDIR
     ninja -C _build install || true
